scripts/requirements.txt
requests
requests_toolbelt
openssh_wrapper
paramiko
proxmoxer
python-dotenv